+++ utils/utils.ts
@@ -1,5 +1,7 @@
 import { DateTime } from 'ptera';
 import { Document, KvValue } from 'kvdex';
+import { dbEntry } from 'db/entry.ts';
+import { dbTag, iTag } from 'db/tag.ts';
 
 export const isURL = (str: string): boolean =>
   (str.length >= 7 && str.substring(0, 7) === 'http://') ||
@@ -56,3 +58,11 @@ export const createDictionaryDocument = <T>(
     acc[doc.id.toString()] = doc.value as T;
     return acc;
   }, {});
+
+export const indexEntries = (entries: dbEntry[], tags: dbTag[]) =>
+  entries.map((entry) => {
+    entry.value.tags = entry.value.tags.map((tag) =>
+      createDictionaryDocument<iTag>(tags)[tag].name
+    );
+    return entry;
+  });